requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2025.1.31
cffi==1.17.1
distro==1.9.0
dj-database-url==2.3.0
Django==4.2.17
django-bootstrap-v5==1.0.11
django-filter==25.1
django-heroku==0.3.1
djangorestframework==3.15.2
factory_boy==3.3.3
Faker==37.0.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
numpy==2.2.4
openai==1.68.0
packaging==24.2
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
whitenoise==6.9.0
